# docs/api.yml
openapi: 3.0.0
servers:
  - url: ...
info:
  version: "latest"
  title: PROG3 - TD2
paths:
  /books:
    get:
      tags:
        - Teams
      summary: get all teams
      operationId: getTeams
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '500':
          description: Internal server error
    post:
      tags:
        - Teams
      summary: create new teams
      operationId: createTeams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeam'
      responses:
        '200':
          description: List of created teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '500':
          description: Internal server error
    put:
      tags :
        - Teams
      summary: update some teams
      operationId: updateTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeam'
      responses:
        '200':
          description: List of updated teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '500':
          description: Internal server error
  /books/{bookId}:
    delete:
      tags :
        - Teams
      summary: delete a specific team
      operationId: deleteTeam
      parameters:
        - in: path
          name: TeamId
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Return the deleted team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '500':
          description: Internal server error
  /players:
    get:
      tags:
        - Players
      summary: get all players
      operationId: getPlayers
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error
    post:
      tags:
        - Players
      summary: create new players
      operationId: CreatePlayers
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreatePlayer'
      responses:
        '200':
          description: List of created players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error
    put:
      tags:
        - Players
      summary: update some players
      operationId: updatePlayers
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UpdatePlayer'
      responses:
        '200':
          description: List of updated players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error
  /players/{playerId}:
    delete:
      tags:
        - Players
      summary: delete a specific player
      operationId: deletePlayer
      parameters:
        - in: path
          name: playerId
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Return the deleted player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error

components:
  schemas:
    Team:
      type: object
      allOf:
        - $ref: '#/components/schemas/UpdateTeam'
    UpdateTeam:
      allOf:
        - $ref: '#/components/schemas/CreateTeam'
      properties:
        id:
          type: integer
    CreateTeam:
      type: object
      properties:
        name:
          type: string
        particularity:
          type: string
    Player:
      type: object
      properties:
        name:
          type: string
      allOf:
        - $ref: '#/components/schemas/UpdatePlayer'
    UpdatePlayer:
      allOf:
        - $ref: '#/components/schemas/CreatePlayer'
      properties:
        id:
          type: integer
    CreatePlayer:
      type: object
      properties:
        name:
          type: string